requirements.txt
astroid==2.1.0
attrs==19.3.0
backcall==0.1.0
betterpath==0.2.2
bleach==3.1.0
Click==7.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.1.7
django-bootstrap3==11.0.0
django-chartjs==1.3
django-crispy-forms==1.7.2
django-filter==2.1.0
django-filters==0.2.1
django-rest-framework==0.1.0
django-tables2==2.0.5
django-widget-tweaks==1.4.3
djangorestframework==3.10.3
entrypoints==0.3
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
l==0.11.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.2.2
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2018.9
pyzmq==18.1.1
qtconsole==4.6.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.3.1
vcversioner==2.16.0.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.1
zipp==0.6.0
zope.interface==4.6.0
